Preliminary multi-notebook app (still a WIP)
