+++ pyproject.toml
@@ -32,6 +32,8 @@ dependencies = [
   "sentence-transformers",
   "stanza",
   "transformers[torch]",
+  "fastapi",
+  "uvicorn",
 ]
 
 [project.optional-dependencies]
@@ -69,8 +71,10 @@ testpaths = ["tests"]
 
 [tool.coverage.run]
 branch = true
+# ignore app notebooks
 omit = [
-  "src/remarx/app/ui.py",   # marimo app notebook
+  "src/remarx/app/corpus_builder.py",
+  "src/remarx/app/quote_finder.py",
 ]
 
 [tool.coverage.report]
@@ -85,7 +89,10 @@ target-version = "py312"
 # Configure src path so ruff import fixes can identify local imports
 src = [ "src" ]
 # Also ignore the app notebook
-extend-exclude = [ "src/remarx/app/ui.py" ]
+extend-exclude = [
+  "src/remarx/app/corpus_builder.py",
+  "src/remarx/app/quote_finder.py",
+]
 
 [tool.ruff.lint]
 # Include these rules in addition to ruff's defaults
