+++ src/remarx/app/__init__.py
@@ -2,6 +2,6 @@
 This module contains the UI and utilities for the remarx app
 """
 
-from remarx.app import ui, utils
+from remarx.app import corpus_builder, quote_finder, utils
 
-__all__ = ["ui", "utils"]
+__all__ = ["corpus_builder", "quote_finder", "utils"]
