+++ src/remarx/app/corpus_builder.py
@@ -23,28 +23,22 @@ def _():
     import tempfile
 
     import remarx
-    from remarx.app.utils import create_temp_input
+    from remarx.app.utils import create_header, create_temp_input
     from remarx.sentence.corpus.create import create_corpus
     from remarx.sentence.corpus import FileInput
-    return FileInput, create_corpus, create_temp_input, mo, pathlib, remarx
-
-
-@app.cell
-def _(mo):
-    mo.vstack(
-        [
-            mo.md("# `remarx`: Quotation Finder").center(),
-            mo.md(
-                "This is the preliminary graphical user interface for the `remarx` software tool."
-            ).center(),
-        ]
+    return (
+        FileInput,
+        create_corpus,
+        create_header,
+        create_temp_input,
+        mo,
+        pathlib,
     )
-    return
 
 
 @app.cell
-def _(mo, remarx):
-    mo.md(rf"""Running `remarx` version: {remarx.__version__}""")
+def _(create_header):
+    create_header()
     return
 
 
@@ -52,7 +46,7 @@ def _(mo, remarx):
 def _(mo):
     mo.md(
         rf"""
-    ## 📝 Sentence Corpus Prep
+    ## 📝 Sentence Corpus Builder
     Create a sentence corpus (`CSV`) from a text.
     This process can be run multiple times for different files (currently one file at a time).
     """
@@ -64,7 +58,7 @@ def _(mo):
 def _(FileInput, mo):
     mo.md(
         rf"""
-    **1. Select Input Text**
+    ### 1. Select Input Text
 
     Upload and select an input file (`{"`, `".join(FileInput.supported_types())}`) for sentence corpus creation.
     Currently, only a single file may be selected.
@@ -99,7 +93,7 @@ def _(mo, select_input):
 def _(mo):
     mo.md(
         r"""
-    **2. Select Output Location**
+    ### 2. Select Output Location
 
     Select the folder where the resulting sentence corpus file should be saved.
     The output CSV file will be named based on the input file.
@@ -147,7 +141,7 @@ def _(mo, select_output_dir):
 def _(mo):
     mo.md(
         r"""
-    **3. Build Sentence Corpus**
+    ### 3. Build Sentence Corpus
 
     Click the "Build Corpus" to run `remarx`.
     The sentence corpus for the input text will be saved as a CSV in the selected save location.
@@ -214,111 +208,6 @@ def _(button, create_corpus, create_temp_input, input_file, mo, output_csv):
         building_msg = f"✅ Sentence corpus saved to: {output_csv}"
 
     mo.md(building_msg).center()
-    return
-
-
-@app.cell
-def _(mo):
-    mo.md(
-        r"""
-    ## 🔍 Quotation Detection
-    Determine and identify the passages of a text corpus (**reuse**) that quote passages from texts in another corpus (**original**).
-    This process requires sentence corpora (`CSVs`) created in the previous section.
-    """
-    )
-    return
-
-
-@app.cell
-def _(mo):
-    mo.md(
-        r"""
-    **1. Select Input CSV Files**
-
-    Browse and select one or more CSV files for each category:
-
-    - **Original Sentence Corpora**: Sentence-level text corpora of the texts that we are searching for quotations of.
-    - **Reuse Sentence Corpora**: Text that may contain quotations from the original text that will be detected.
-    """
-    )
-    return
-
-
-@app.cell
-def _(mo, pathlib):
-    # Create file browsers for quotation detection (CSV files only)
-    original_csv_browser = mo.ui.file_browser(
-        selection_mode="file",
-        multiple=True,
-        initial_path=pathlib.Path.home(),
-        filetypes=[".csv"],
-    )
-
-    reuse_csv_browser = mo.ui.file_browser(
-        selection_mode="file",
-        multiple=True,
-        initial_path=pathlib.Path.home(),
-        filetypes=[".csv"],
-    )
-    return original_csv_browser, reuse_csv_browser
-
-
-@app.cell
-def _(mo, original_csv_browser, reuse_csv_browser):
-    # Process file selections for quotation detection
-    original_csvs = original_csv_browser.value or []
-    reuse_csvs = reuse_csv_browser.value or []
-
-    original_msg = (
-        f"{len(original_csvs)} files selected"
-        if original_csvs
-        else "No original text files selected"
-    )
-    reuse_msg = (
-        f"{len(reuse_csvs)} files selected"
-        if reuse_csvs
-        else "No reuse text files selected"
-    )
-
-    original_callout_type = "success" if original_csvs else "warn"
-    reuse_callout_type = "success" if reuse_csvs else "warn"
-
-    # Create side-by-side file browser interface
-    mo.hstack(
-        [
-            mo.callout(
-                mo.vstack(
-                    [
-                        mo.md(
-                            "**🗂 Select Original Sentence Corpora (CSVs)**"
-                        ).center(),
-                        original_csv_browser,
-                        mo.md(original_msg),
-                    ]
-                ),
-                kind=original_callout_type,
-            ),
-            mo.callout(
-                mo.vstack(
-                    [
-                        mo.md(
-                            "**♻️ Select Reuse Sentence Corpora (CSVs)**"
-                        ).center(),
-                        reuse_csv_browser,
-                        mo.md(reuse_msg),
-                    ]
-                ),
-                kind=reuse_callout_type,
-            ),
-        ],
-        widths="equal",
-        gap=1.2,
-    )
-    return
-
-
-@app.cell
-def _():
     return
